requirements.txt
Flask
paho-mqtt
pymongo
dnspython
gunicorn
matplotlib